analyzer/requirements.txt
addict
accelerate
calendar
datetime
diffusers
diffusers[torch]
fairscale
gradio
huggingface_hub
hashlib
litellm
matplotlib
nltk
numpy
onnxruntime
opencv_python
pandas
pathlib
Pillow
pycocotools
PyYAML
requests
safetensors
seaborn
setuptools
scipy
scikit-learn
supervision
sys
tensorboard
tensorflow
termcolor
timm
torch
torchvision
tqdm
transformers
wget
